// src/generated/nebula/vector.ts
export interface VectorCommonMetaData {
  numRecords: number;
  vectorContentType: number;
}
// ...
export interface VectorTableMetaData {
  tableType: number;
  numRecords: bigint;
  rowType?: RowType | undefined;
  numBatches: number;
  /** unit: minutes */
  timeZoneOffset: number;
  /** true: small endian, false: big endian */
  isLittleEndian: boolean;
  graphSchema: PropertyGraphSchema[];
}

export interface NestedVector {
  numNestedVectors: number;
  commonMetaData?: VectorCommonMetaData | undefined;
  specialMetaData: Buffer;
  vectorData: Buffer;
  nullBitMap: Buffer;
  nestedVectors: NestedVector[];
}

export interface VectorBatch {
  vectors: NestedVector[];
}

export interface VectorResultTable {
  dataLayoutVersion: Buffer;
  meta?: VectorTableMetaData | undefined;
  batch: VectorBatch[];
}
// ...
export const VectorCommonMetaData: MessageFns<VectorCommonMetaData> = {
  encode(message: VectorCommonMetaData, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.numRecords !== 0) {
      writer.uint32(8).uint32(message.numRecords);
    }
    if (message.vectorContentType !== 0) {
      writer.uint32(16).uint32(message.vectorContentType);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): VectorCommonMetaData {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseVectorCommonMetaData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.numRecords = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.vectorContentType = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): VectorCommonMetaData {
    return {
      numRecords: isSet(object.numRecords) ? globalThis.Number(object.numRecords) : 0,
      vectorContentType: isSet(object.vectorContentType) ? globalThis.Number(object.vectorContentType) : 0,
    };
  },

  toJSON(message: VectorCommonMetaData): unknown {
    const obj: any = {};
    if (message.numRecords !== 0) {
      obj.numRecords = Math.round(message.numRecords);
    }
    if (message.vectorContentType !== 0) {
      obj.vectorContentType = Math.round(message.vectorContentType);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<VectorCommonMetaData>, I>>(base?: I): VectorCommonMetaData {
    return VectorCommonMetaData.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<VectorCommonMetaData>, I>>(object: I): VectorCommonMetaData {
    const message = createBaseVectorCommonMetaData();
    message.numRecords = object.numRecords ?? 0;
    message.vectorContentType = object.vectorContentType ?? 0;
    return message;
  },
};
// ...
export const NestedVector: MessageFns<NestedVector> = {
  encode(message: NestedVector, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.numNestedVectors !== 0) {
      writer.uint32(8).uint32(message.numNestedVectors);
    }
    if (message.commonMetaData !== undefined) {
      VectorCommonMetaData.encode(message.commonMetaData, writer.uint32(18).fork()).join();
    }
    if (message.specialMetaData.length !== 0) {
      writer.uint32(26).bytes(message.specialMetaData);
    }
    if (message.vectorData.length !== 0) {
      writer.uint32(34).bytes(message.vectorData);
    }
    if (message.nullBitMap.length !== 0) {
      writer.uint32(42).bytes(message.nullBitMap);
    }
    for (const v of message.nestedVectors) {
      NestedVector.encode(v!, writer.uint32(50).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): NestedVector {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseNestedVector();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.numNestedVectors = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.commonMetaData = VectorCommonMetaData.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.specialMetaData = Buffer.from(reader.bytes());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.vectorData = Buffer.from(reader.bytes());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.nullBitMap = Buffer.from(reader.bytes());
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.nestedVectors.push(NestedVector.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): NestedVector {
    return {
      numNestedVectors: isSet(object.numNestedVectors) ? globalThis.Number(object.numNestedVectors) : 0,
      commonMetaData: isSet(object.commonMetaData) ? VectorCommonMetaData.fromJSON(object.commonMetaData) : undefined,
      specialMetaData: isSet(object.specialMetaData)
        ? Buffer.from(bytesFromBase64(object.specialMetaData))
        : Buffer.alloc(0),
      vectorData: isSet(object.vectorData) ? Buffer.from(bytesFromBase64(object.vectorData)) : Buffer.alloc(0),
      nullBitMap: isSet(object.nullBitMap) ? Buffer.from(bytesFromBase64(object.nullBitMap)) : Buffer.alloc(0),
      nestedVectors: globalThis.Array.isArray(object?.nestedVectors)
        ? object.nestedVectors.map((e: any) => NestedVector.fromJSON(e))
        : [],
    };
  },

  toJSON(message: NestedVector): unknown {
    const obj: any = {};
    if (message.numNestedVectors !== 0) {
      obj.numNestedVectors = Math.round(message.numNestedVectors);
    }
    if (message.commonMetaData !== undefined) {
      obj.commonMetaData = VectorCommonMetaData.toJSON(message.commonMetaData);
    }
    if (message.specialMetaData.length !== 0) {
      obj.specialMetaData = base64FromBytes(message.specialMetaData);
    }
    if (message.vectorData.length !== 0) {
      obj.vectorData = base64FromBytes(message.vectorData);
    }
    if (message.nullBitMap.length !== 0) {
      obj.nullBitMap = base64FromBytes(message.nullBitMap);
    }
    if (message.nestedVectors?.length) {
      obj.nestedVectors = message.nestedVectors.map((e) => NestedVector.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<NestedVector>, I>>(base?: I): NestedVector {
    return NestedVector.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<NestedVector>, I>>(object: I): NestedVector {
    const message = createBaseNestedVector();
    message.numNestedVectors = object.numNestedVectors ?? 0;
    message.commonMetaData = (object.commonMetaData !== undefined && object.commonMetaData !== null)
      ? VectorCommonMetaData.fromPartial(object.commonMetaData)
      : undefined;
    message.specialMetaData = object.specialMetaData ?? Buffer.alloc(0);
    message.vectorData = object.vectorData ?? Buffer.alloc(0);
    message.nullBitMap = object.nullBitMap ?? Buffer.alloc(0);
    message.nestedVectors = object.nestedVectors?.map((e) => NestedVector.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const VectorResultTable: MessageFns<VectorResultTable> = {
  encode(message: VectorResultTable, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.dataLayoutVersion.length !== 0) {
      writer.uint32(10).bytes(message.dataLayoutVersion);
    }
    if (message.meta !== undefined) {
      VectorTableMetaData.encode(message.meta, writer.uint32(18).fork()).join();
    }
    for (const v of message.batch) {
      VectorBatch.encode(v!, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): VectorResultTable {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseVectorResultTable();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.dataLayoutVersion = Buffer.from(reader.bytes());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.meta = VectorTableMetaData.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.batch.push(VectorBatch.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): VectorResultTable {
    return {
      dataLayoutVersion: isSet(object.dataLayoutVersion)
        ? Buffer.from(bytesFromBase64(object.dataLayoutVersion))
        : Buffer.alloc(0),
      meta: isSet(object.meta) ? VectorTableMetaData.fromJSON(object.meta) : undefined,
      batch: globalThis.Array.isArray(object?.batch) ? object.batch.map((e: any) => VectorBatch.fromJSON(e)) : [],
    };
  },

  toJSON(message: VectorResultTable): unknown {
    const obj: any = {};
    if (message.dataLayoutVersion.length !== 0) {
      obj.dataLayoutVersion = base64FromBytes(message.dataLayoutVersion);
    }
    if (message.meta !== undefined) {
      obj.meta = VectorTableMetaData.toJSON(message.meta);
    }
    if (message.batch?.length) {
      obj.batch = message.batch.map((e) => VectorBatch.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<VectorResultTable>, I>>(base?: I): VectorResultTable {
    return VectorResultTable.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<VectorResultTable>, I>>(object: I): VectorResultTable {
    const message = createBaseVectorResultTable();
    message.dataLayoutVersion = object.dataLayoutVersion ?? Buffer.alloc(0);
    message.meta = (object.meta !== undefined && object.meta !== null)
      ? VectorTableMetaData.fromPartial(object.meta)
      : undefined;
    message.batch = object.batch?.map((e) => VectorBatch.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}